disparos

+++ Project_8_Solution/Source/ModuleParticles.cpp
@@ -451,7 +451,7 @@ Update_Status ModuleParticles::PostUpdate()
 	return Update_Status::UPDATE_CONTINUE;
 }
 
-Particle* ModuleParticles::AddParticle(const Particle& particle, int x, int y, float speedx, float speedy,Collider::Type colliderType, uint delay)
+Particle* ModuleParticles::AddParticle(const Particle& particle, float x, float y, float speedx, float speedy,Collider::Type colliderType, uint delay)
 {
 	Particle* p = new Particle(particle);
